+++ crates/pg/src/lib.rs
@@ -2,8 +2,10 @@ use async_trait::async_trait;
 use bb8::{Pool, PooledConnection};
 use bb8_postgres::PostgresConnectionManager;
 use sqls::{
-    select_account_profiles_by_db_cr_accounts, select_approvers,
-    select_rule_instance_by_type_role_account, select_rule_instance_by_type_role_state,
+    profile::{select_account_profiles_by_db_cr_accounts, select_approvers},
+    rule_instance::{
+        select_rule_instance_by_type_role_account, select_rule_instance_by_type_role_state,
+    },
 };
 use std::{env, error::Error, sync::Arc};
 use tokio_postgres::{types::ToSql, NoTls, Row};
